% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{cellClass}
\alias{cellClass}
\alias{cellClass,SPData-method}
\alias{cellClass<-}
\alias{cellClass<-,SPData,numeric-method}
\alias{cellClass<-,SPData-method}
\title{Classes of each cell}
\usage{
cellClass(object)

cellClass(object) <- value

\S4method{cellClass}{SPData}(object)
}
\arguments{
\item{object}{The SPData object to use}

\item{value}{A new vector of cell classes. Must be the same length as the number of cells.}
}
\value{
A numeric vector of cell classes
}
\description{
If cells within a given tissue have been classified into different
types then get/set the classes.

Classes of each cell
}

