% Generated by roxygen2 (4.0.1): do not edit by hand
\name{channelCorr}
\alias{channelCorr}
\title{Correlation plot of channels within the sample}
\usage{
channelCorr(sp, cell.class = NULL)
}
\arguments{
\item{sp}{The \code{SPData} object to di}

\item{cell.class}{A specific class of cells to pick out for correlation plots}
}
\description{
Correlation plot of channels within the sample
}

