% Generated by roxygen2 (4.0.1): do not edit by hand
\name{channelPlot}
\alias{channelPlot}
\title{Boxplots the distributions of channels depending on their class}
\usage{
channelPlot(sp, channel.ids)
}
\arguments{
\item{sp}{The \code{SPData} object to use}

\item{channel.ids}{A numeric vector of channel indicies}
}
\description{
Boxplots the distributions of channels depending on their class
}

