% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[[}
\alias{[[}
\alias{[[,SPExp,numeric-method}
\alias{[[,SPExp-method}
\title{Access a sample}
\usage{
\S4method{[[}{SPExp,numeric}(x, i)
}
\arguments{
\item{x}{The SPExp instance to use}

\item{i}{The index of the \code{SPData} object within the \code{SPExp} to extract}
}
\value{
The \code{SPData} object in slot \code{i} of the \code{SPExp} instance
}
\description{
Extracts a single sample from a \code{SPExp} object
}

