% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[}
\alias{[}
\alias{[,SPData,ANY,ANY,ANY-method}
\alias{[,SPData-method}
\title{Subset an SPData set}
\usage{
\S4method{[}{SPData,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{i}{Cells to subset}

\item{j}{Channels to subset}

\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }
}
\value{
An SPData object reduced to cells \code{i} and channels \code{j}
}
\description{
Select SPData[i,j] for cells \code{i} and channels \code{j}.
Note that this does not subset out nearest neighbours also.
}
\examples{
\dontrun{
## subset to cells 1,3,5 and channels 8 to 10:
i <- c(1,3,5)
j <- 8:10
sp.reduced <- sp[i,j]
}
}

