% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateNeighbourfromID}
\alias{generateNeighbourfromID}
\title{Generate nearest neighbour data}
\usage{
generateNeighbourfromID(sp)
}
\arguments{
\item{sp}{The \code{SPData} object to use}
}
\value{
A \code{SPData} object with the neighbour data regenerated
}
\description{
Some processing may be performed on the cell-by-channel matrix and the nearest neighbour
data needs updated. Note that the class architecture means that when the neighbours of a cell
are requested, it is not generated fresh from the cell matrix but retreived from a separate copy
specifically for neighbour data. Therefore, if any transformations are applied to the neighbour
data, you will need to regenerate them using this function.
}

