% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ID}
\alias{ID}
\alias{ID,SPData-method}
\alias{ID<-}
\alias{ID<-,SPData,numeric-method}
\alias{ID<-,SPData-method}
\title{Get and set the sample ID}
\usage{
ID(object)

ID(object) <- value

\S4method{ID}{SPData}(object)
}
\arguments{
\item{object}{The SPData object for which to set the ID.}

\item{value}{The new sample ID}
}
\description{
Get the sample ID.

Set the sample ID.
}

