% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,SPExp-method}
\alias{initialize,SPExp-method}
\title{Initialize SPExp}
\usage{
\S4method{initialize}{SPExp}(.Object, dir, files, spdata)
}
\arguments{
\item{dir}{The \code{SPExp} directory}

\item{files}{A list of the original experiment filenames}

\item{spdata}{A list of \code{SPData} objects}

\item{.Object}{ An object:  see the Details section.}
}
\description{
Initialize SPExp
}

