% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neighbourCorr}
\alias{neighbourCorr}
\title{Correlation plot of channels with neighbours}
\usage{
neighbourCorr(sp, cell.class = NULL)
}
\arguments{
\item{sp}{The \code{SPData} object to di}

\item{cell.class}{A specific class of cells to pick out for neighbour correlation plots}
}
\description{
Correlation plot of channels with neighbours
}

