% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{neighbourIDs}
\alias{neighbourIDs}
\alias{neighbourIDs,SPData-method}
\title{Nearest neighbour IDs}
\usage{
neighbourIDs(object)

\S4method{neighbourIDs}{SPData}(object)
}
\arguments{
\item{object}{The \code{SPData} object from which to extract the neighbour IDs.}
}
\description{
A list of length \emph{n} for \emph{n} cells, where the \emph{i}th
entry is a vector of length \emph{m} that holds the IDs of the nearest
neighbour cells of cell \emph{i}, if cell \emph{i} has \emph{m} nearest
neighbours.
}

