% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,SPExp-method}
\alias{show,SPExp-method}
\title{Displays content of \code{SPExp} object}
\usage{
\S4method{show}{SPExp}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Displays content of \code{SPExp} object
}

