% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SpatialPRo-package}
\alias{SpatialPRo}
\alias{SpatialPRo-package}
\title{Methods for spatial proteomics data}
\description{
SpatialPRo: Methods for spatial proteomics data.
}
\details{
SpatialPRo can store, normalise and cluster spatially
resolved single-cell proteomics data.
}
\author{
Kieran Campbell \email{kieran.campbell@dpag.ox.ac.uk}
}

