% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[}
\alias{[}
\alias{[,SPExp-method}
\title{Subset a \code{SPExp} object.}
\usage{
\S4method{[}{SPExp,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{The \code{SPExp} instance to subset}

\item{i}{The samples to retain}
}
\value{
A subsetted \code{SPExp} object
}
\description{
Returns a new \code{SPExp} object subsetted using the first \code{i} instances
}

