% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[[}
\alias{[[}
\alias{[[<-,SPExp,numeric,missing,SPData-method}
\alias{[[<-,SPExp-method}
\title{Sets a single sample in a \code{SPExp} object}
\arguments{
\item{x}{The SPExp instance to use}

\item{i}{The index of the \code{SPData} object within the \code{SPExp} to set}

\item{value}{An object of class \code{SPData} to replace slot \code{i}}
}
\description{
Sets a single sample in a \code{SPExp} object
}

