% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SPExperiment}
\alias{SPExperiment}
\title{Create a new \code{SPExp} object}
\usage{
SPExperiment(dir, files, spdata)
}
\arguments{
\item{dir}{Experiment directory}

\item{files}{Experiment file names}

\item{spdata}{The list of \code{SPData} objects}
}
\description{
Create a new \code{SPExp} object
}

