% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{xy}
\alias{xy}
\alias{xy,SPData-method}
\alias{xy<-}
\alias{xy<-,SPData,matrix-method}
\alias{xy<-,SPData-method}
\title{2D Cell Coordinates}
\usage{
xy(object)

xy(object) <- value

\S4method{xy}{SPData}(object)
}
\arguments{
\item{object}{The SPData object from which to retrieve ethe coordinates}

\item{value}{New cell coordinanes. Must be an ncell-by-2 matrix.}
}
\description{
A matrix of size \emph{n} by 2, for \emph{n} cells. The first
column is the \emph{x} coordinate in the tissue, and the
second is the \emph{y}.

Set the cell coordinates.
}

